feat: updated user registration functionality with cloudinary image uploads and api error handling.

+++ src/utils/cloudinary.js
@@ -7,19 +7,18 @@ cloudinary.config({
   api_secret: process.env.CLOUDINARY_API_SECRET,
 });
 
-const uploadCloudinary = async (localFilePath) => {
+const uploadOnCloudinary = async (localFilePath) => {
   try {
     if (!localFilePath) return null;
     const response = await cloudinary.uploader.upload(localFilePath, {
       resource_type: "auto",
     });
-    console.log("File upload success!!");
-    console.log("🚀 ~ uploadCloudinary ~ response:", response);
+    fs.unlinkSync(localFilePath);
     return response;
   } catch (error) {
-    fs.unlinkSync(localFilePath);
+    fs.existsSync(localFilePath) && fs.unlinkSync(localFilePath);
     return null;
   }
 };
 
-export { uploadCloudinary };
+export { uploadOnCloudinary };
